[Chenyu Zhao] : optimize controllers directory structure
